feat: implement password encoder
implement password encoder
fix typo in task module folders
