+++ src/main/java/com/projects/taskmanager/task/domain/ports/ITaskRepositoryPort.java
@@ -1,0 +1,6 @@
+package com.projects.taskmanager.task.domain.ports;
+
+import com.projects.taskmanager.core.domain.ports.IRepositoryPort;
+import com.projects.taskmanager.task.domain.models.Task;
+
+public interface ITaskRepositoryPort extends IRepositoryPort<Task, String> {}
