+++ src/main/java/com/projects/taskmanager/task/infrastructure/controllers/TasksController.java
@@ -1,0 +1,17 @@
+package com.projects.taskmanager.task.infrastructure.controllers;
+
+import com.projects.taskmanager.core.infrastructure.controllers.CoreController;
+import com.projects.taskmanager.task.adapters.controllers.TaskControllerAdapter;
+import com.projects.taskmanager.task.domain.models.Task;
+import com.projects.taskmanager.task.infrastructure.dtos.TaskDto;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/task")
+public class TasksController extends CoreController<Task, TaskDto, String> {
+
+  public TasksController(TaskControllerAdapter adapter) {
+    super(adapter);
+  }
+}
