+++ src/main/java/com/projects/taskmanager/task/infrastructure/repositories/entities/TaskEntity.java
@@ -1,0 +1,25 @@
+package com.projects.taskmanager.task.infrastructure.repositories.entities;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+import java.time.LocalDate;
+
+@Data
+@Entity
+@Table(name = "tasks")
+public class TaskEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.UUID)
+  private String id;
+
+  private String name;
+
+  private LocalDate date;
+
+  private boolean reminder;
+
+  @Column(name = "user_id", updatable = false, nullable = false)
+  private String userId;
+}
