--- src/main/java/com/projects/taskmanager/tasks/domain/ports/ITaskRepositoryPort.java
@@ -1,6 +1,0 @@
-package com.projects.taskmanager.tasks.domain.ports;
-
-import com.projects.taskmanager.core.domain.ports.IRepositoryPort;
-import com.projects.taskmanager.tasks.domain.models.Task;
-
-public interface ITaskRepositoryPort extends IRepositoryPort<Task, String> {}
