--- src/main/java/com/projects/taskmanager/tasks/infrastructure/repositories/TaskRepository.java
@@ -1,8 +1,0 @@
-package com.projects.taskmanager.tasks.infrastructure.repositories;
-
-import com.projects.taskmanager.core.infrastructure.repositories.IRepository;
-import com.projects.taskmanager.tasks.infrastructure.repositories.entities.TaskEntity;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TaskRepository extends IRepository<TaskEntity, String> {}
