--- src/main/java/com/projects/taskmanager/tasks/infrastructure/repositories/entities/TaskEntity.java
@@ -1,18 +1,0 @@
-package com.projects.taskmanager.tasks.infrastructure.repositories.entities;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-@Entity
-@Table(name = "tasks")
-public class TaskEntity {
-  @Id
-  @GeneratedValue(strategy = GenerationType.UUID)
-  private String id;
-  private String name;
-  private LocalDate date;
-  private boolean reminder;
-}
