+++ src/main/java/com/projects/taskmanager/user/infrastructure/repositories/entities/UserEntity.java
@@ -36,6 +36,10 @@ public class UserEntity {
   private boolean credentialsNoExpired;
 
   @ManyToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
-  @JoinTable(name = "user_roles", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"))
+  @JoinTable(
+    name = "user_roles",
+    joinColumns = @JoinColumn(name = "user_id"),
+    inverseJoinColumns = @JoinColumn(name = "role_id")
+  )
   private Set<RoleEntity> roles = new HashSet<>();
 }
